requirements.txt
docopt
pystache
PyYAML
nose
mock
fig
